+++ src/components/PayPalScriptProvider.tsx
@@ -19,16 +19,16 @@ It manages state for script loading so children components like `<PayPalButtons 
 Note: You always should use this component as a wrapper for  `PayPalButtons`, `PayPalMarks`, `PayPalMessages` and `BraintreePayPalButtons` components.
  */
 export const PayPalScriptProvider: FC<ScriptProviderProps> = ({
-    options = { "client-id": "test" },
+    options = { clientId: "test" },
     children,
     deferLoading = false,
 }: ScriptProviderProps) => {
     const [state, dispatch] = useReducer(scriptReducer, {
         options: {
-            ...options,
-            [SCRIPT_ID]: `${getScriptID(options)}`,
             [SDK_SETTINGS.DATA_SDK_INTEGRATION_SOURCE]:
                 SDK_SETTINGS.DATA_SDK_INTEGRATION_SOURCE_VALUE,
+            ...options,
+            [SCRIPT_ID]: `${getScriptID(options)}`,
         },
         loadingStatus: deferLoading
             ? SCRIPT_LOADING_STATE.INITIAL
@@ -66,7 +66,10 @@ export const PayPalScriptProvider: FC<ScriptProviderProps> = ({
                 if (isSubscribed) {
                     dispatch({
                         type: DISPATCH_ACTION.LOADING_STATUS,
-                        value: SCRIPT_LOADING_STATE.REJECTED,
+                        value: {
+                            state: SCRIPT_LOADING_STATE.REJECTED,
+                            message: String(err),
+                        },
                     });
                 }
             });