+++ src/stories/braintree/BraintreePayPalButtons.stories.tsx
@@ -27,6 +27,7 @@ import type { FC } from "react";
 import type {
     PayPalScriptOptions,
     PayPalButtonsComponentOptions,
+    FUNDING_SOURCE,
 } from "@paypal/paypal-js";
 import type { StoryFn } from "@storybook/react";
 import type { DocsContextProps } from "@storybook/addon-docs";
@@ -38,7 +39,7 @@ import type {
 
 type StoryProps = {
     style: PayPalButtonsComponentOptions["style"];
-    fundingSource: string;
+    fundingSource: FUNDING_SOURCE;
     disabled: boolean;
     amount: string;
     currency: string;
@@ -46,7 +47,7 @@ type StoryProps = {
 
 const uid = generateRandomString();
 const scriptProviderOptions: PayPalScriptOptions = {
-    "client-id": "test",
+    clientId: "test",
     components: "buttons",
     ...getOptionsFromQueryString(),
 };
@@ -78,6 +79,8 @@ export default {
         createBillingAgreement: { table: { category: COMPONENT_EVENTS } },
         createSubscription: { table: { category: COMPONENT_EVENTS } },
         onShippingChange: { table: { category: COMPONENT_EVENTS } },
+        onShippingAddressChange: { table: { category: COMPONENT_EVENTS } },
+        onShippingOptionsChange: { table: { category: COMPONENT_EVENTS } },
         onApprove: { table: { category: COMPONENT_EVENTS } },
         onCancel: { table: { category: COMPONENT_EVENTS } },
         onClick: { table: { category: COMPONENT_EVENTS } },
@@ -125,9 +128,9 @@ export default {
                             <PayPalScriptProvider
                                 options={{
                                     ...scriptProviderOptions,
-                                    "data-client-token": clientToken,
-                                    "data-namespace": uid,
-                                    "data-uid": uid,
+                                    dataClientToken: clientToken,
+                                    dataNamespace: uid,
+                                    dataUid: uid,
                                     intent: isBillingAggrement
                                         ? "tokenize"
                                         : "capture",