+++ src/stories/payPalHostedFields/PayPalHostedFields.stories.tsx
@@ -11,7 +11,9 @@ import {
     getOptionsFromQueryString,
     generateRandomString,
     getClientToken,
-    HEROKU_SERVER,
+    FLY_SERVER,
+    CREATE_ORDER_URL,
+    CAPTURE_ORDER_URL,
 } from "../utils";
 import {
     COMPONENT_PROPS_CATEGORY,
@@ -36,29 +38,18 @@ type StoryProps = {
 };
 
 const uid = generateRandomString();
-const TOKEN_URL = `${HEROKU_SERVER}/api/paypal/hosted-fields/auth`;
-const CREATE_ORDER_URL = `${HEROKU_SERVER}/api/paypal/checkout/orders`;
+const TOKEN_URL = `${FLY_SERVER}/api/paypal/generate-client-token`;
 const RED_COLOR = "#dc3545";
 const GREEN_COLOR = "#28a745";
 const scriptProviderOptions: PayPalScriptOptions = {
-    "client-id":
-        "AdOu-W3GPkrfuTbJNuW9dWVijxvhaXHFIRuKrLDwu14UDwTTHWMFkUwuu9D8I1MAQluERl9cFOd7Mfqe",
+    clientId:
+        "AduyjUJ0A7urUcWtGCTjanhRBSzOSn9_GKUzxWDnf51YaV1eZNA0ZAFhebIV_Eq-daemeI7dH05KjLWm",
     components: "buttons,hosted-fields",
     ...getOptionsFromQueryString(),
 };
 const CREATE_ORDER = "createOrder";
 const SUBMIT_FORM = "submitForm";
 const CAPTURE_ORDER = "captureOrder";
-
-/**
- * Get dynamically the capture order URL to fetch the payment info
- *
- * @param orderId the order identifier
- * @returns an URL string
- */
-function captureOrderUrl(orderId: string): string {
-    return `${HEROKU_SERVER}/api/paypal/checkout/orders/${orderId}/capture`;
-}
 
 /**
  * Functional component to submit the hosted fields form
@@ -95,8 +86,12 @@ const SubmitPayment = ({ customStyle }: { customStyle?: CSSProperties }) => {
                 action(CAPTURE_ORDER)(
                     `Sending ${ORDER_ID} to custom endpoint to capture the payment information`
                 );
-                fetch(captureOrderUrl(data.orderId), {
-                    method: "post",
+                fetch(CAPTURE_ORDER_URL, {
+                    method: "POST",
+                    headers: {
+                        "Content-Type": "application/json",
+                    },
+                    body: JSON.stringify({ orderID: data.orderId }),
                 })
                     .then((response) => response.json())
                     .then((data) => {
@@ -233,9 +228,9 @@ export default {
                             <PayPalScriptProvider
                                 options={{
                                     ...scriptProviderOptions,
-                                    "data-client-token": clientToken,
-                                    "data-namespace": uid,
-                                    "data-uid": uid,
+                                    dataClientToken: clientToken,
+                                    dataNamespace: uid,
+                                    dataUid: uid,
                                 }}
                             >
                                 <Story />
@@ -256,20 +251,17 @@ export const Default: FC<StoryProps> = ({ amount, styles, style }) => {
                     "Start creating the order in custom endpoint"
                 );
                 return fetch(CREATE_ORDER_URL, {
-                    method: "post",
+                    method: "POST",
                     headers: {
                         "Content-Type": "application/json",
                     },
                     body: JSON.stringify({
-                        purchase_units: [
+                        cart: [
                             {
-                                amount: {
-                                    value: amount,
-                                    currency_code: "USD",
-                                },
+                                sku: "1blwyeo8",
+                                quantity: 2,
                             },
                         ],
-                        intent: "CAPTURE",
                     }),
                 })
                     .then((response) => response.json())
@@ -336,7 +328,7 @@ export const ExpirationDate: FC<{ amount: string }> = ({ amount }) => {
         <PayPalHostedFieldsProvider
             createOrder={() =>
                 fetch(CREATE_ORDER_URL, {
-                    method: "post",
+                    method: "POST",
                     headers: {
                         "Content-Type": "application/json",
                     },