+++ src/stories/utils.ts
@@ -3,15 +3,19 @@ import format from "string-template";
 import { SDK_QUERY_KEYS } from "@paypal/sdk-constants/dist/module";
 
 import type { ReactNode } from "react";
+import type { OrderResponseBody } from "@paypal/paypal-js";
 
 // FIXME: problem with union on key
 type TokenResponse = {
     [key in "clientToken" | "client_token"]: string;
 } & { success?: boolean };
 
-export const HEROKU_SERVER = "https://braintree-sdk-demo.herokuapp.com";
-const CLIENT_TOKEN_URL = `${HEROKU_SERVER}/api/braintree/auth`;
-const SALE_URL = `${HEROKU_SERVER}/api/braintree/sale`;
+export const FLY_SERVER = "https://react-paypal-js-storybook.fly.dev";
+export const CREATE_ORDER_URL = `${FLY_SERVER}/api/paypal/create-order`;
+export const CAPTURE_ORDER_URL = `${FLY_SERVER}/api/paypal/capture-order`;
+
+const CLIENT_TOKEN_URL = `${FLY_SERVER}/api/braintree/generate-client-token`;
+const SALE_URL = `${FLY_SERVER}/api/braintree/sale`;
 const allowedSDKQueryParams = Object.keys(SDK_QUERY_KEYS).map(
     (key) => SDK_QUERY_KEYS[key]
 );
@@ -33,7 +37,9 @@ export function getOptionsFromQueryString(): Record<string, string> {
 }
 
 export async function getClientToken(url = CLIENT_TOKEN_URL): Promise<string> {
-    const response: TokenResponse = await (await fetch(url)).json();
+    const response: TokenResponse = await (
+        await fetch(url, { method: "POST" })
+    ).json();
 
     return response?.client_token || response?.clientToken;
 }
@@ -72,4 +78,32 @@ export function generateFundingSource(fundingSource?: string): string {
     return `fundingSource=${
         fundingSource ? `"${fundingSource}"` : "{undefined}"
     }`;
+}
+
+export function createOrder(
+    cart: { sku: string; quantity: number }[]
+): Promise<OrderResponseBody> {
+    return fetch(CREATE_ORDER_URL, {
+        method: "POST",
+        headers: {
+            "Content-Type": "application/json",
+        },
+        // use the "body" param to optionally pass additional order information
+        // like product ids and quantities
+        body: JSON.stringify({ cart }),
+    }).then((response) => response.json());
+}
+
+export function onApprove(data: {
+    orderID: string;
+}): Promise<OrderResponseBody> {
+    return fetch(CAPTURE_ORDER_URL, {
+        method: "POST",
+        headers: {
+            "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+            orderID: data.orderID,
+        }),
+    }).then((response) => response.json());
 }