+++ src/types/scriptProviderTypes.ts
@@ -1,4 +1,3 @@
-import { SCRIPT_ID } from "../constants";
 import { BraintreePayPalCheckout } from "./braintree/paypalCheckout";
 import { DISPATCH_ACTION, SCRIPT_LOADING_STATE } from "./enums";
 
@@ -6,7 +5,7 @@ import type { ReactNode, Dispatch } from "react";
 import type { PayPalScriptOptions } from "@paypal/paypal-js";
 
 export interface ReactPayPalScriptOptions extends PayPalScriptOptions {
-    [SCRIPT_ID]: string;
+    [key: string]: any;
 }
 
 export type ScriptReducerAction =
@@ -15,8 +14,12 @@ export type ScriptReducerAction =
           value: SCRIPT_LOADING_STATE;
       }
     | {
+          type: `${DISPATCH_ACTION.LOADING_STATUS}`;
+          value: { state: SCRIPT_LOADING_STATE; message: string };
+      }
+    | {
           type: `${DISPATCH_ACTION.RESET_OPTIONS}`;
-          value: PayPalScriptOptions | ReactPayPalScriptOptions;
+          value: ReactPayPalScriptOptions;
       }
     | {
           type: `${DISPATCH_ACTION.SET_BRAINTREE_INSTANCE}`;
@@ -31,6 +34,7 @@ export type InitialState = {
 export interface ScriptContextState {
     options: ReactPayPalScriptOptions;
     loadingStatus: SCRIPT_LOADING_STATE;
+    loadingStatusErrorMessage?: string;
     braintreePayPalCheckoutInstance?: BraintreePayPalCheckout;
     dispatch?: Dispatch<ScriptReducerAction>;
 }
@@ -44,7 +48,7 @@ export interface ScriptContextDerivedState
 }
 
 export interface ScriptProviderProps {
-    options: PayPalScriptOptions;
+    options: ReactPayPalScriptOptions;
     children?: ReactNode;
     deferLoading?: boolean;
 }